+++ device-minion-runtime/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added a new SysCall to fliush all L1 and L2 with a single thread
+- [SW-18547] Added support for custom U-mode kernel stack configuration.
 ### Changed
 - [SW-18472] Reduced log message length for bus error
 ### Deprecated